Projeto Final
Finalizando o projeto, criando a opção de encerrar a aplicação, e re-organizando o código respeitando as boas práticas do Clean Code, para termos maior facilidade para futuras implementações e manutenções.
