--- bytebank_Arrays/test
@@ -1,1 +1,0 @@
-s
